+++ egui/src/style.rs
@@ -138,6 +138,9 @@ pub struct Visuals {
     /// The color used for `Hyperlink`,
     pub hyperlink_color: Color32,
 
+    /// Background color behind code-styled monospaced labels.
+    pub code_bg_color: Color32,
+
     pub window_corner_radius: f32,
     pub window_shadow: Shadow,
 
@@ -165,6 +168,14 @@ impl Visuals {
     pub fn text_color(&self) -> Color32 {
         self.override_text_color
             .unwrap_or_else(|| self.widgets.noninteractive.text_color())
+    }
+
+    pub fn weak_text_color(&self) -> Color32 {
+        self.widgets.disabled.text_color()
+    }
+
+    pub fn strong_text_color(&self) -> Color32 {
+        self.widgets.active.text_color()
     }
 }
 
@@ -283,6 +294,7 @@ impl Default for Visuals {
             selection: Default::default(),
             dark_bg_color: Color32::from_gray(10),
             hyperlink_color: Color32::from_rgb(90, 170, 255),
+            code_bg_color: Color32::from_gray(64),
             window_corner_radius: 10.0,
             window_shadow: Shadow::big(),
             resize_corner_size: 12.0,
@@ -500,6 +512,7 @@ impl Visuals {
             selection,
             dark_bg_color,
             hyperlink_color,
+            code_bg_color,
             window_corner_radius,
             window_shadow,
             resize_corner_size,
@@ -514,6 +527,7 @@ impl Visuals {
         ui.collapsing("selection", |ui| selection.ui(ui));
         ui_color(ui, dark_bg_color, "dark_bg_color");
         ui_color(ui, hyperlink_color, "hyperlink_color");
+        ui_color(ui, code_bg_color, "code_bg_color");
         ui.add(Slider::f32(window_corner_radius, 0.0..=20.0).text("window_corner_radius"));
         shadow_ui(ui, window_shadow, "Window shadow:");
         ui.add(Slider::f32(resize_corner_size, 0.0..=20.0).text("resize_corner_size"));